Draft (1/2): Implementing Mixing of streams
This supports mixing 2+ streams, one of which we need to be microphone data and the other sound data.

In order for this to happen, the console must send a stream to the server, and the server must load an mp3 (currently wav because of no encoder), consequently mix, and finally send to every requesting client.

Have also implemented a bit of communication.

All data inside the streams must be raw PCM data.  (This is also the reason why we do not use MediaRecorder for microphone recording.)

This code is untested.