Fix bug where if microphone_stream was empty it wouldn't play the music
Why we want this?
Say the producer hasn't yet enabled the mic and is just streaming music.  We need to support this.
It works now.

Also, fix bug where the stream would start and then immediately stop.  It was a problem of variable garbage collection.  Making everything global fixed it, it seems.